requirements.txt
anthropic>=0.40.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
